refactor(todo): centralized todo phase validation
- Moved todo status, item, phase, and runtime validation into a cycle-free shared module.
- Re-exported the contract from the todo tool while making eval persistence and session tracking consume the canonical validator directly.

Fixes #10647
